+++ package.json
@@ -39,7 +39,7 @@
     "react-popper": "^2.2.5",
     "react-toastify": "^8.2.0",
     "react-virtual": "^2.10.4",
-    "mdsmirror": "^0.1.13",
+    "mdsmirror": "^0.1.16",
     "styled-components": "^5.3.5",
     "tailwindcss": "^3.0.24",
     "zustand": "^3.7.2"
